Separate compatibility tests
