+++ compat/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "enderpy-compat"
+version = "0.0.0"
+authors = { workspace = true }
+edition = { workspace = true }
+rust-version = { workspace = true }
+homepage = { workspace = true }
+documentation = { workspace = true }
+repository = { workspace = true }
+license = { workspace = true }
+
+[dependencies]
+serde_json = "1.0"
+serde = "1.0"
+miette = "5.5"
+enderpy_python_parser = { path = "../parser"}
+tabled = "0.15"
+terminal_size = "0.3"
+assert-json-diff = "2.0"
+pretty_assertions = "1.4"
+which = "6.0.1"
+reqwest = { version = "0.11", features = ["blocking"] }
+zip = "0.6"
