+++ parser/Cargo.toml
@@ -20,7 +20,6 @@ unicode-id-start = "1.0.3"
 miette = { version = "5.6.0", features = ["fancy"] }
 thiserror = "1.0.40"
 is-macro = "0.3.5"
-which = "6.0.1"
 
 [dev-dependencies]
 codspeed-criterion-compat.workspace = true
@@ -28,10 +27,6 @@ criterion.workspace = true
 insta.workspace = true
 reqwest = { version= "0.12.4", features = ["blocking"] }
 tokio.workspace = true
-tabled = "0.15"
-terminal_size = "0.3"
-assert-json-diff = "2.0"
-pretty_assertions = "1.4"
 
 [lib]
 bench = false
