+++ parser/src/lib.rs
@@ -1,12 +1,11 @@
 mod lexer;
-mod parser;
+pub mod parser;
 
 pub use crate::{
     lexer::Lexer,
     parser::{ast, parser::Parser},
 };
 pub mod error;
-pub mod runpython;
 pub mod token;
 
 pub fn get_row_col_position(start: u32, end: u32, line_starts: &[u32]) -> (u32, u32, u32, u32) {
